pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mediated-coevo"
version = "0.1.0"
description = "Context-Aware Mediation for Multi-Agent Skill Co-Evolution"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "litellm>=1.70.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tomli-w>=1.0.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
medcoevo = "mediated_coevo.main:app"

[tool.setuptools]
packages = {find = {where = ["src"], include = ["mediated_coevo*"]}}
